requirements.txt
aiohttp==3.13.2
aiohttp_socks==0.10.2
fastapi==0.123.5
httpx==0.28.1
pydantic==2.12.5
requests==2.32.5
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.5.2
py-clob-client>=0.4.0
py-builder-signing-sdk>=0.1.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
lxml>=5.0.0